refactor stage and actor management, move helper methods to helper.* files, remove stage.cpp and ActorContainer.h
