+++ actors/pythonactor.c
@@ -1,4 +1,5 @@
 #include "pythonactor.h"
+#include "helpers.h"
 #include "pyzmsg.h"
 #include <wchar.h>
 #include "config.h"
@@ -9,7 +10,6 @@
 #endif
 #ifdef __UTYPE_LINUX
 #include <sys/inotify.h>
-#include "config.h"
 
 zmsg_t *
 s_pythonactor_set_file(pythonactor_t *self, const char *filename);
@@ -122,43 +122,6 @@ s_init_inotify(pythonactor_t *self, const char *filename, sphactor_actor_t *acto
 }
 #endif
 
-// https://stackoverflow.com/questions/2736753/how-to-remove-extension-from-file-name
-static char *
-s_remove_ext(const char* myStr) {
-    char *retStr;
-    char *lastExt;
-    if (myStr == NULL) return NULL;
-    if ((retStr = malloc (strlen (myStr) + 1)) == NULL) return NULL;
-    strcpy (retStr, myStr);
-    lastExt = strrchr (retStr, '.');
-    if (lastExt != NULL)
-        *lastExt = '\0';
-    return retStr;
-}
-
-static char *
-s_basename(char const *path)
-{
-    char *s = strrchr(path, '/');
-    if (!s)
-        return strdup(path);
-    else
-        return strdup(s + 1);
-}
-
-static bool
-s_dir_exists(const wchar_t *pypath)
-{
-    // bleeh we need to convert to char as czmq only supports char paths
-    char path[PATH_MAX];
-    wcstombs(path, pypath, PATH_MAX);
-    zfile_t *dir = zfile_new(NULL, path);
-    assert(dir);
-    bool ret = zfile_is_directory(dir);
-    zfile_destroy(&dir);
-    return ret;
-}
-
 static const char *pythonactorcapabilities =
         "capabilities\n"
         "    data\n"
@@ -178,6 +141,105 @@ static const char *pythonactorcapabilities =
         "    output\n"
         //TODO: Perhaps add NatNet output type so we can filter the data multiple times...
         "        type = \"OSC\"\n";
+
+static PyObject *
+s_py_zosc_tuple(pythonactor_t *self, zosc_t *oscmsg)
+{
+    assert(self);
+    assert(oscmsg);
+    const char *format = zosc_format(oscmsg);
+
+    PyObject *rettuple = PyTuple_New((Py_ssize_t) strlen(format) );
+
+    char type = '0';
+    Py_ssize_t pos = 0;
+    const void *data =  zosc_first(oscmsg, &type);
+    while(data)
+    {
+        PyObject *o = NULL;
+        switch (type)
+        {
+        case('i'):
+        {
+           int32_t val = 9;
+           int rc = zosc_pop_int32(oscmsg, &val);
+           assert(rc == 0);
+           o = PyLong_FromLong((long)val);
+           break;
+        }
+        case('h'):
+        {
+           int64_t val = 0;
+           int rc = zosc_pop_int64(oscmsg, &val);
+           assert(rc == 0);
+           o = PyLong_FromLong((long)val);
+           break;
+        }
+        case('f'):
+        {
+           float flt_v = 0.f;
+           int rc = zosc_pop_float(oscmsg, &flt_v);
+           assert(rc == 0);
+           o = PyFloat_FromDouble((double)flt_v);
+           break;
+        }
+        case 'd':
+        {
+           double dbl_v = 0.;
+           int rc = zosc_pop_double(oscmsg, &dbl_v);
+           assert(rc == 0);
+           o = PyFloat_FromDouble(dbl_v);
+           break;
+        }
+        case 's':
+        {
+           char *str;
+           int rc = zosc_pop_string(oscmsg, &str);
+           assert(rc == 0);
+           o = PyUnicode_Decode(str, strlen(str), "ascii", "ignore");
+           zstr_free(&str);
+           break;
+        }
+        case 'c':
+        {
+           char chr;
+           int rc = zosc_pop_char(oscmsg, &chr);
+           assert(rc == 0);
+           o = Py_BuildValue("C", chr);
+           break;
+        }
+        case 'F':
+        case 'T':
+        {
+           bool bl;
+           int rc = zosc_pop_bool(oscmsg, &bl);
+           assert(rc == 0);
+           if (bl)
+               o = Py_True;
+           else
+               o = Py_False;
+           Py_INCREF(o);
+           break;
+        }
+        default:
+           assert(0);
+        }
+
+        if (o == NULL && PyErr_Occurred())
+        {
+           PyErr_Print();
+        }
+        else
+        {
+           int rc = PyTuple_SetItem(rettuple, pos, o);
+           assert(rc == 0);
+        }
+
+        data = zosc_next(oscmsg, &type);
+        pos++;
+    }
+    return rettuple;
+}
 
 void
 s_py_set_timeout(pythonactor_t *self, sphactor_event_t *ev)
@@ -313,214 +375,6 @@ s_pythonactor_set_file(pythonactor_t *self, const char *filename)
     return NULL;
 }
 
-static zosc_t *
-s_py_zosc(PyObject *pAddress, PyObject *pData)
-{
-    assert( PyUnicode_Check(pAddress) );
-    assert( PyList_Check(pData) );
-    PyObject *stringbytes = PyUnicode_AsASCIIString(pAddress);
-    zosc_t *ret = zosc_new( PyBytes_AsString( stringbytes) );
-    Py_DECREF(stringbytes);
-
-    // iterate
-    for ( Py_ssize_t i=0;i<PyList_Size(pData);++i )
-    {
-        PyObject *item = PyList_GetItem(pData, i);
-        assert(item);
-        // determine type, first check if boolean otherwise it will be an int
-        if (PyBool_Check(item))
-        {
-            long b = PyLong_AsLong(item);
-            if (b)
-                zosc_append(ret, "T");
-            else
-                zosc_append(ret, "F");
-        }
-        else if ( PyLong_Check(item) )
-        {
-            long v = PyLong_AsLong(item);
-            zosc_append(ret, "h", v);
-        }
-        else if (PyFloat_Check(item))
-        {
-            double v = PyFloat_AsDouble(item);
-            zosc_append(ret, "d", v);
-        }
-        else if (PyUnicode_Check(item))
-        {
-            PyObject* ascii = PyUnicode_AsASCIIString(item);
-            assert(ascii);
-            char* s = PyBytes_AsString(ascii);
-            zosc_append(ret, "s", s);
-            Py_DECREF(ascii);
-        }
-        else if (PyBytes_Check(item)) // this can be used to force 32bit int, ie: struct.pack("I", 32)
-        {
-            Py_ssize_t bytesize = PyBytes_Size(item);
-            if ( bytesize == 4 ) // 32bit int
-            {
-                char * buffer = PyBytes_AsString(item);
-                uint32_t* vp = (uint32_t *)(buffer);
-                zosc_append(ret, "i", *vp);
-            }
-            else
-                zsys_warning("I don't know what to do with these bytes, please help: https://github.com/hku-ect/gazebosc");
-        }
-        else
-        {
-            // not a supported native python type try ctypes
-            PyTypeObject* type = Py_TYPE(item);
-            const char * typename = _PyType_Name(type);
-            if (streq(typename, "c_int"))
-            {
-                zsys_warning("Unsupported ctypes.c_int until we find a way to access the value :S");
-            }
-            else
-                zsys_warning("unsupported python type");
-        }
-
-    }
-
-    return ret;
-}
-
-static PyObject *
-s_py_zosc_tuple(pythonactor_t *self, zosc_t *oscmsg)
-{
-    assert(self);
-    assert(oscmsg);
-    const char *format = zosc_format(oscmsg);
-
-    PyObject *rettuple = PyTuple_New((Py_ssize_t) strlen(format) );
-
-    char type = '0';
-    Py_ssize_t pos = 0;
-    const void *data =  zosc_first(oscmsg, &type);
-    while(data)
-    {
-        PyObject *o = NULL;
-        switch (type)
-        {
-        case('i'):
-        {
-            int32_t val = 9;
-            int rc = zosc_pop_int32(oscmsg, &val);
-            assert(rc == 0);
-            o = PyLong_FromLong((long)val);
-            break;
-        }
-        case('h'):
-        {
-            int64_t val = 0;
-            int rc = zosc_pop_int64(oscmsg, &val);
-            assert(rc == 0);
-            o = PyLong_FromLong((long)val);
-            break;
-        }
-        case('f'):
-        {
-            float flt_v = 0.f;
-            int rc = zosc_pop_float(oscmsg, &flt_v);
-            assert(rc == 0);
-            o = PyFloat_FromDouble((double)flt_v);
-            break;
-        }
-        case 'd':
-        {
-            double dbl_v = 0.;
-            int rc = zosc_pop_double(oscmsg, &dbl_v);
-            assert(rc == 0);
-            o = PyFloat_FromDouble(dbl_v);
-            break;
-        }
-        case 's':
-        {
-            char *str;
-            int rc = zosc_pop_string(oscmsg, &str);
-            assert(rc == 0);
-            o = PyUnicode_Decode(str, strlen(str), "ascii", "ignore");
-            zstr_free(&str);
-            break;
-        }
-        case 'c':
-        {
-            char chr;
-            int rc = zosc_pop_char(oscmsg, &chr);
-            assert(rc == 0);
-            o = Py_BuildValue("C", chr);
-            break;
-        }
-        case 'F':
-        case 'T':
-        {
-            bool bl;
-            int rc = zosc_pop_bool(oscmsg, &bl);
-            assert(rc == 0);
-            if (bl)
-                o = Py_True;
-            else
-                o = Py_False;
-            Py_INCREF(o);
-            break;
-        }
-        default:
-            assert(0);
-        }
-
-        if (o == NULL && PyErr_Occurred())
-        {
-            PyErr_Print();
-        }
-        else
-        {
-            int rc = PyTuple_SetItem(rettuple, pos, o);
-            assert(rc == 0);
-        }
-
-        data = zosc_next(oscmsg, &type);
-        pos++;
-    }
-    return rettuple;
-}
-
-void python_add_path(const char *path)
-{
-    PyGILState_STATE gstate;
-    gstate = PyGILState_Ensure();
-    char *script = (char*)malloc(512 * sizeof(char));
-    // Prints "Hello world!" on hello_world
-    snprintf(script, 512 * sizeof(char),
-             "import sys\n"
-             "if \"%s\" not in sys.path:\n"
-             "\tsys.path.append(r\"%s/site-packages\")\n"
-             "print(\"added path %s/site-packages to sys.path\")\n",
-             path, path, path);
-    int rc = PyRun_SimpleString(script);
-    assert(rc == 0);
-    PyGILState_Release(gstate);
-    free(script);
-}
-
-void python_remove_path(const char *path)
-{
-    PyGILState_STATE gstate;
-    gstate = PyGILState_Ensure();
-    char *script = (char*)malloc(512 * sizeof(char));
-    // Prints "Hello world!" on hello_world
-    snprintf(script, 512 * sizeof(char),
-             "import sys\n"
-             "try:\n"
-             "\tsys.path.remove(r\"%s/site-packages\")\n"
-             "except Exception:\n"
-             "\tpass\n"
-             "print(\"removed path %s/site-packages from sys.path\")\n",
-             path, path);
-    int rc = PyRun_SimpleString(script);
-    assert(rc == 0);
-    PyGILState_Release(gstate);
-    free(script);
-}
-
 int python_init()
 {
     Py_UnbufferedStdioFlag = 1;
@@ -611,58 +465,6 @@ int python_init()
 
     sphactor_register("Python", &pythonactor_handler, zconfig_str_load(pythonactorcapabilities), &pythonactor_new_helper, NULL); // https://stackoverflow.com/questions/65957511/typedef-for-a-registering-a-constructor-function-in-c
     return rc;
-}
-
-PyObject *python_call_file_func(const char *file, const char *func, const char *fmt, ...)
-{
-    PyObject *pName, *pModule, *pFunc, *pValue;
-    pValue = NULL;
-    pName = PyUnicode_DecodeFSDefault(file);
-    /* Error checking of pName left out */
-    assert(pName);
-    pModule = PyImport_Import(pName);
-    Py_DECREF(pName);
-
-    if (pModule != NULL)
-    {
-        pFunc = PyObject_GetAttrString(pModule, func);
-        /* pFunc is a new reference */
-
-        if (pFunc && PyCallable_Check(pFunc)) {
-
-            va_list argsptr;
-            va_start(argsptr, fmt);
-            PyObject *pArgs = Py_VaBuildValue(fmt, argsptr);
-            assert(pArgs);
-            pValue = PyObject_CallObject(pFunc, pArgs);
-            //vprintf(fmt, argsptr);
-            va_end(argsptr);
-            if (pValue != NULL) {
-                printf("Result of call: %ld\n", PyLong_AsLong(pValue));
-                Py_DECREF(pValue);
-            }
-            else {
-                Py_DECREF(pFunc);
-                Py_DECREF(pModule);
-                PyErr_Print();
-                fprintf(stderr,"Call failed\n");
-            }
-        }
-        else {
-            if (PyErr_Occurred())
-                PyErr_Print();
-            fprintf(stderr, "Cannot find function \"%s\"\n", func);
-        }
-        Py_XDECREF(pFunc);
-        Py_DECREF(pModule);
-    }
-    else
-    {
-        PyErr_Print();
-        fprintf(stderr, "Failed to load \"%s\"\n", file);
-    }
-    return pValue;
-
 }
 
 // this is needed because we have to conform to returning a void * thus we need to wrap
